// vedio-catalog/video-catalog/src/models/Movie.ts
export interface Movie {
    imdbID: string;
    Title: string;
    Year: string;
    Poster: string;
    Type: string;
  }

  export interface MovieDetailsType {
    Title: string;
    Year: string;
    Genre: string;
    Director: string;
    Actors: string;
    Plot: string;
    Poster: string;
    imdbRating: string;
  }
  
  